Fix GPRELEASE-21: clear old release files from trunk.
In fact, this isn't a real fix for GPRELEASE-21 since the release files are still in trunk. But removing the old ones is a significant improvement and I think this is the most we're going to do on this issue since we will be moving away from Subversion for plugin repositories.