Improve error reporting when plugin publication fails.
The plugin was simply reporting a status code and not much else. In fact, it was reporting a 302 for 'permission denied' leaving users bemused. With the latest changes to the website, the users are now told explicitly whether the authentication failed, they didn't have permission to publish, or something else happened.