Handle missing credentials when publishing to grailsCentral.
Unlike when publishing to a Subversion repo or a Maven one, the user is never
asked for their credentials when deploying to Grails Central if those creds
are not configured in settings.groovy (or BuildConfig.groovy). This commit
fixes that by asking for the username and password if the initial auth fails.