Check whether publication was successful.
Previously, the plugin treated a 200 response code as a success, but since the
server was deploying the plugin asynchronously this was an unreliable guide to
actual success.

The new web site now has an API endpoint that allows you to check the status of
a plugin publication, so we now poll that for a few seconds until we get a
SUCCESS, FAIL, or time out. This means the user gets much better feedback on
whether the plugin was actually published or not.